Use size_t when iterating through a string

+++ project4/array.cpp
@@ -47,7 +47,7 @@ int countVowels(const string array[], int n) {
         int num_vowels = 0;
 
         for (int i = 0; i < n; i++) {
-            for (int j = 0; j < array[i].length(); j++) {
+            for (size_t j = 0; j < array[i].length(); j++) {
                 const char letter = tolower(array[i].at(j));
                 num_vowels = letter == 'a' || letter == 'e' || letter == 'i' ||
                                      letter == 'o' || letter == 'u' ||
